requirements.txt
# GitHub Security Scanner — Dependencias
# Instalar con: pip install -r requirements.txt

# HTTP client
requests>=2.31.0

# Variables de entorno
python-dotenv>=1.0.0

# Reglas personalizadas en YAML
PyYAML>=6.0.1

# CLI coloreado (opcional, para mejores logs)
colorama>=0.4.6; sys_platform == "win32"
